restruct project structure
