+++ .gitignore
@@ -89,3 +89,7 @@ ENV/
 
 # Rope project settings
 .ropeproject
+
+# Ignore some dataset dir
+data/
+parsed-data/
